fee_sys/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect2==0.6.0
Django==3.2.13
django-addresses==0.1
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
django-invoice==0.3.2
django-relativedelta==1.1.2
future==0.18.2
html5lib==1.1
idna==3.3
jsons==1.6.2
lxml==4.9.0
oscrypto==1.3.0
Pillow==9.1.1
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
typish==1.9.3
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.7
